requirements.txt
numpy==1.20.3
opencv_python==4.5.3.56
Pillow==8.3.2